Refactor kitty key handling and align naming
Split kitty-specific input and tty logic into input-kitty.c and tty-kitty.c, and wire them through tmux.h and Makefile.am.

Replace per-screen kitty stack array with current/saved two-slot state and update CSI push/pop/set/query handling accordingly.

Adopt Sp-/Hy- modifier prefixes, drop RealMeta parsing/emission and kitty bit 0x20 mapping, and standardize terminology to VT10x/extended and function key wording.

Expand regress/kitty-keys.sh expectations for the single saved kitty level behavior and updated terminology.
