+++ Makefile.am
@@ -172,6 +172,7 @@ dist_tmux_SOURCES = \
 	hooks.c \
 	hyperlinks.c \
 	input-keys.c \
+	input-kitty.c \
 	input.c \
 	job.c \
 	key-bindings.c \
@@ -211,6 +212,7 @@ dist_tmux_SOURCES = \
 	tty-acs.c \
 	tty-draw.c \
 	tty-features.c \
+	tty-kitty.c \
 	tty-keys.c \
 	tty-term.c \
 	tty.c \
