+++ key-string.c
@@ -296,20 +296,15 @@ key_string_get_modifiers(const char **string)
 
 	modifiers = 0;
 	while ((*string)[0] != '\0') {
-		/* Check long-form modifier prefixes first. */
-		if (strncasecmp(*string, "Super-", 6) == 0) {
+		/* Check multicharacter modifier prefixes first. */
+		if (strncasecmp(*string, "Sp-", 3) == 0) {
 			modifiers |= KEYC_SUPER;
-			*string += 6;
+			*string += 3;
 			continue;
 		}
-		if (strncasecmp(*string, "Hyper-", 6) == 0) {
+		if (strncasecmp(*string, "Hy-", 3) == 0) {
 			modifiers |= KEYC_HYPER;
-			*string += 6;
-			continue;
-		}
-		if (strncasecmp(*string, "RealMeta-", 9) == 0) {
-			modifiers |= KEYC_META_REAL;
-			*string += 9;
+			*string += 3;
 			continue;
 		}
 		/* Single-char prefix: X- where X is C/M/S. */
@@ -443,11 +438,9 @@ key_string_lookup_key(key_code key, int with_flags)
 
 	/* Fill in the modifiers. */
 	if (key & KEYC_SUPER)
-		strlcat(out, "Super-", sizeof out);
+		strlcat(out, "Sp-", sizeof out);
 	if (key & KEYC_HYPER)
-		strlcat(out, "Hyper-", sizeof out);
-	if (key & KEYC_META_REAL)
-		strlcat(out, "RealMeta-", sizeof out);
+		strlcat(out, "Hy-", sizeof out);
 	if (key & KEYC_CTRL)
 		strlcat(out, "C-", sizeof out);
 	if (key & KEYC_META)
